# Conflicts:
#	browser/Reynard/Client/Interface/Addons/AddonErrorPresentation.swift
#	browser/Reynard/Client/Interface/Addons/AddonPermissionPromptViewController.swift
#	browser/Reynard/Client/Interface/Addons/AddonPermissionSupport.swift
#	browser/Reynard/Client/Interface/Addons/AddonUpdateCoordinator.swift
#	browser/Reynard/Client/Interface/Chrome/AddressBar/AddressBar.swift
#	browser/Reynard/Client/Interface/ContentView/WebContent/PermissionPrompt/PermissionPromptPresenter.swift
#	browser/Reynard/Client/Interface/ContentView/WebContent/PromptPresenter.swift
#	browser/Reynard/Client/Interface/Homepage/Sections/FrequentlyVisited/FrequentlyVisitedSiteActions.swift
#	browser/Reynard/Client/Interface/Homepage/Sections/PrivateBrowsing/PrivateBrowsingSectionViewController.swift
#	browser/Reynard/Client/Interface/Homepage/Sections/RecentlyClosedTabs/RecentlyCloseTabItemActions.swift
#	browser/Reynard/Client/Interface/Homepage/Sections/Recommendations/PerformanceRecommendationViewController.swift
#	browser/Reynard/Client/Interface/Homepage/Sections/UpdateAvailable/UpdateAvailableViewController.swift
#	browser/Reynard/Client/Interface/Library/Bookmarks/BookmarksViewController.swift
#	browser/Reynard/Client/Interface/Library/Bookmarks/EditBookmark/BookmarkFolderRowCell.swift
#	browser/Reynard/Client/Interface/Library/ClearData/ClearDataTimeframe.swift
#	browser/Reynard/Client/Interface/Library/ClearData/ClearDownloadsViewController.swift
#	browser/Reynard/Client/Interface/Library/ClearData/ClearHistoryViewController.swift
#	browser/Reynard/Client/Interface/Library/Downloads/DownloadSection.swift
#	browser/Reynard/Client/Interface/Library/Downloads/DownloadsViewController.swift
#	browser/Reynard/Client/Interface/Library/Settings/Sections/About/AboutSettingsSection.swift
#	browser/Reynard/Client/Interface/Library/Settings/Sections/General/Addons/AddonDetails/AddonDetailsPreferencesViewController.swift
#	browser/Reynard/Client/Interface/Library/Settings/Sections/General/Addons/AddonInformation/AddonInformationPreferencesViewController.swift
#	browser/Reynard/Client/Interface/Library/Settings/Sections/General/Addons/AddonPermissions/AddonPermissionsPreferencesViewController.swift
#	browser/Reynard/Client/Interface/Library/Settings/Sections/General/Addons/AddonsPreferencesViewController.swift
#	browser/Reynard/Client/Interface/Library/Settings/Sections/General/Appearance/AppearancePreferencesViewController.swift
#	browser/Reynard/Client/Interface/Library/Settings/Sections/General/Appearance/Sections/AddressBarPosition/AddressBarPositionPickerCell.swift
#	browser/Reynard/Client/Interface/Library/Settings/Sections/General/Appearance/Sections/AppAppearance/AppAppearancePickerCell.swift
#	browser/Reynard/Client/Interface/Library/Settings/Sections/General/Appearance/Sections/WebsiteDisplay/PageZoom/PageZoomPreferencesViewController.swift
#	browser/Reynard/Client/Interface/Library/Settings/Sections/General/Browsing/BrowsingPreferencesViewController.swift
#	browser/Reynard/Client/Interface/Library/Settings/Sections/General/Compatibility/CompatibilityPreferencesViewController.swift
#	browser/Reynard/Client/Interface/Library/Settings/Sections/General/Compatibility/UserAgentOverrides/UserAgentOverridesPreferencesViewController.swift
#	browser/Reynard/Client/Interface/Library/Settings/Sections/General/GeneralSettingsSection.swift
#	browser/Reynard/Client/Interface/Library/Settings/Sections/General/Homepage/HomepagePreferencesViewController.swift
#	browser/Reynard/Client/Interface/Library/Settings/Sections/General/Homepage/HomepageSection/HomepageSectionPreferencesViewController.swift
#	browser/Reynard/Client/Interface/Library/Settings/Sections/General/NewTab/NewTabPreferencesViewController.swift
#	browser/Reynard/Client/Interface/Library/Settings/Sections/General/Search/SearchEngine/SearchEnginePreferencesViewController.swift
#	browser/Reynard/Client/Interface/Library/Settings/Sections/JIT/JITSettingsSection.swift
#	browser/Reynard/Client/Interface/Library/Settings/Sections/Privacy/ClearBrowsingData/ClearBrowsingDataViewController.swift
#	browser/Reynard/Client/Interface/Library/Settings/Sections/Privacy/PrivacySettingsSection.swift
#	browser/Reynard/Client/Interface/Library/Settings/Sections/Privacy/SitePermissions/SitePermissionDetails/SitePermissionDetailsViewController.swift
#	browser/Reynard/Client/Interface/Library/Settings/Sections/Privacy/SitePermissions/SitePermissionsViewController.swift
#	browser/Reynard/Client/Interface/Library/Settings/Sections/Updates/UpdatesSettingsSection.swift
#	browser/Reynard/Client/Interface/Search/SearchViewController.swift
#	browser/Reynard/Client/Interface/SiteSettings/SiteSettingsUtils.swift
#	browser/Reynard/Client/Interface/SiteSettings/SiteSettingsViewController.swift
#	browser/Reynard/Client/Interface/TabOverview/TabOverviewCard.swift
#	browser/Reynard/Client/Interface/TabOverview/TabOverviewCollection.swift
#	browser/Reynard/Client/Interface/TabOverview/TabOverviewToolbar/TabOverviewBottomToolbar.swift
#	browser/Reynard/Client/Interface/TabOverview/TabOverviewToolbar/TabOverviewTopToolbar.swift
#	browser/Reynard/Client/SessionManagement/Coordinators/PermissionCoordinator.swift
#	browser/Reynard/JIT/Interface/JITFailure.swift
#	browser/Reynard/JIT/JITController.swift
#	browser/Scripts/AddGecko.sh

+++ browser/Reynard/Client/Stores/DownloadStore.swift
@@ -347,27 +347,30 @@ final class DownloadStore: NSObject {
         }
     }
     
-    func clearCompletedDownloads(since startDate: Date?) {
+    func clearCompletedDownloadFiles(since startDate: Date? = nil) {
         stateQueue.async {
+            let removedDownloads: [PersistedDownloadEntry]
             if let startDate {
+                removedDownloads = self.persistedDownloads.filter { $0.addedAt >= startDate }
                 self.persistedDownloads.removeAll { $0.addedAt >= startDate }
             } else {
+                removedDownloads = self.persistedDownloads
                 self.persistedDownloads.removeAll()
             }
             
-            self.savePersistedDownloadsLocked()
-            self.postDidChange()
-        }
-    }
-    
-    func clearCompletedDownloadFiles() {
-        stateQueue.async {
-            let fileURLs = (try? self.fileManager.contentsOfDirectory(
-                at: self.storage.downloadsDirectoryURL,
-                includingPropertiesForKeys: nil
-            )) ?? []
+            let fileURLs: [URL]
+            if startDate == nil {
+                fileURLs = (try? self.fileManager.contentsOfDirectory(
+                    at: self.storage.downloadsDirectoryURL,
+                    includingPropertiesForKeys: nil
+                )) ?? []
+            } else {
+                fileURLs = removedDownloads.map {
+                    self.storage.downloadsDirectoryURL.appendingPathComponent($0.relativePath, isDirectory: false)
+                }
+            }
             
-            for fileURL in fileURLs {
+            for fileURL in Set(fileURLs) {
                 try? self.fileManager.removeItem(at: fileURL)
             }
             
@@ -379,6 +382,10 @@ final class DownloadStore: NSObject {
             }
             
             for active in self.activeDownloads.values {
+                if let startDate, active.addedAt < startDate {
+                    continue
+                }
+                
                 if self.fileManager.fileExists(atPath: active.destinationURL.path) {
                     try? self.fileManager.removeItem(at: active.destinationURL)
                 }
@@ -555,7 +562,7 @@ final class DownloadStore: NSObject {
     // MARK: - Files
     
     private func resolvedFileName(suggestedFileName: String?, sourceURL: URL, mimeType: String?) -> String {
-        let fallbackName = sourceURL.lastPathComponent.isEmpty ? "Download" : sourceURL.lastPathComponent
+        let fallbackName = sourceURL.lastPathComponent.isEmpty ? NSLocalizedString("Download", comment: "") : sourceURL.lastPathComponent
         let initialName = sanitizeFileName(suggestedFileName ?? fallbackName)
         
         guard URL(fileURLWithPath: initialName).pathExtension.isEmpty,
@@ -583,7 +590,7 @@ final class DownloadStore: NSObject {
             .filter { !$0.isEmpty }
             .joined(separator: "-")
         
-        return sanitized.isEmpty ? "Download" : sanitized
+        return sanitized.isEmpty ? NSLocalizedString("Download", comment: "") : sanitized
     }
     
     private func makeUniqueDestinationURLLocked(for fileName: String) -> URL {
